erp5_core: define property of propertysheets by changing directly the reference
Until now, it was only possible to define the ID on properties, then the
reference was guessed from the ID. We even had constraint to make sure
when we define a standard "foo" property that the ID was "foo_property".
Then the reference "foo" was guessed for the ID.

This is not user friendly, it is more logical to make the ID non editable
and guess it from the reference. Therefore now the developer could define
the reference, like "foo", and the corresponding ID will be automatically
set.